Add a REST API for downloading files
- Improve file server core components to be more flexible and changeable.
- Use a dedicated connection to download a single file. (Close the connection after the download is complete.)
- Add a RestAssured test.

+++ src/main/java/practice/netty/message/FileChunkTxResponse.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class FileChunkTxResponse implements Message {
     private final boolean endOfFile;
     private final long start;
-    private final long length;
+    private final int length;
     private final String filePath;
 
     @Override
